# Add performance benchmark for indexselect OP 