--- a/benchmarks/cpp/nvfuser/CMakeLists.txt
+++ b/benchmarks/cpp/nvfuser/CMakeLists.txt
@@ -23,6 +23,7 @@ if(USE_CUDA)
     transpose.cpp
     matmul.cpp
     timm.cpp
+    indexselect.cpp
     utils.cpp
     main.cpp)