As the backend/frontend PRs for adding `index_select` op are merged, this PR is opened to add the corresponding benchmark for `index_select` op. 


----
Moving from https://github.com/chang-l/pytorch-nvfuser/pull/1 


## Benchmarks

### Benchmarking the setup/schedule/lower/compile overheads
- cmd line:
  `./build/bin/nvfuser_bench  --benchmark_filter=IndexSelect_`
- Sample output:
```
-------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations
-------------------------------------------------------------------
IndexSelect_SetupFusion        20.9 us         20.9 us         6705
IndexSelect_AutoSchedule       7477 us         7474 us           18
IndexSelect_Lower              14.0 ms         14.0 ms            9
IndexSelect_Compile             153 ms          153 ms            1
IndexSelect_RunFusion          33.4 us         33.4 us         4170
```
### Benchmarking a single `index_select` operator (not fused with other operations) together with torch `at::index_select`
- cmd line:
`./build/bin/nvfuser_bench  --benchmark_filter=IndexSelectSimple`
- Sample output:
```
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                                                            Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
NvFuserScheduler_IndexSelectSimple_Outer_fp32___GRAPH/NvFuserScheduler_IndexSelectSimple_Outer_fp32/32768/65536/manual_time       99.6 us          154 us         1398 bytes_per_second=316.057G/s 2D Schedule at 1/ Split block into y-dim/Vectorize, Factor: 2/Launch_Parameters[block(1/2/64)/grid(1/32768/1)/0]
Baseline_IndexSelectSimple_Outer_fp32/32768/65536/manual_time                                                                      135 us          178 us         1035 bytes_per_second=233.582G/s

```
### Benchmarking `index_select`  fused with elem-wise operators (addition and multiplication) by varying the extent of input and select axis
- cmd line:
`./build/bin/nvfuser_bench  --benchmark_filter=IndexSelect_Outer`
- Sample output:
```
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                                                          Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
NvFuserScheduler_IndexSelect_Outer_fp32___GRAPH/NvFuserScheduler_IndexSelectFused_Outer_fp32/128/16/manual_time            3.84 us         36.3 us        35497 bytes_per_second=5.96971G/s 1D//Launch_Parameters[block(1/1/128)/grid(1/1/16)/0]
NvFuserScheduler_IndexSelect_Outer_fp32___GRAPH/NvFuserScheduler_IndexSelectFused_Outer_fp32/512/16/manual_time            3.82 us         36.4 us        36607 bytes_per_second=6.00123G/s 1D//Launch_Parameters[block(1/1/128)/grid(1/1/16)/0]
NvFuserScheduler_IndexSelect_Outer_fp32___GRAPH/NvFuserScheduler_IndexSelectFused_Outer_fp32/4096/16/manual_time           3.85 us         35.8 us        36403 bytes_per_second=5.96461G/s 1D//Launch_Parameters[block(1/1/128)/grid(1/1/16)/0]
NvFuserScheduler_IndexSelect_Outer_fp32___GRAPH/NvFuserScheduler_IndexSelectFused_Outer_fp32/32768/16/manual_time          3.83 us         36.3 us        36765 bytes_per_second=5.99809G/s 1D//Launch_Parameters[block(1/1/128)/grid(1/1/16)/0]
...
NvFuserScheduler_IndexSelect_Outer_fp32___GRAPH/NvFuserScheduler_IndexSelectFused_Outer_fp32/128/32768/manual_time         53.3 us         85.6 us         2629 bytes_per_second=882.254G/s 2D Schedule at 1/ Split block into y-dim/Vectorize, Factor: 2/Launch_Parameters[block(1/2/64)/grid(1/16384/1)/0]
NvFuserScheduler_IndexSelect_Outer_fp32___GRAPH/NvFuserScheduler_IndexSelectFused_Outer_fp32/512/32768/manual_time         53.4 us         85.7 us         2621 bytes_per_second=879.847G/s 2D Schedule at 1/ Split block into y-dim/Vectorize, Factor: 2/Launch_Parameters[block(1/2/64)/grid(1/16384/1)/0]
NvFuserScheduler_IndexSelect_Outer_fp32___GRAPH/NvFuserScheduler_IndexSelectFused_Outer_fp32/4096/32768/manual_time        60.3 us         92.6 us         2321 bytes_per_second=778.851G/s 2D Schedule at 1/ Split block into y-dim/Vectorize, Factor: 2/Launch_Parameters[block(1/2/64)/grid(1/16384/1)/0]
NvFuserScheduler_IndexSelect_Outer_fp32___GRAPH/NvFuserScheduler_IndexSelectFused_Outer_fp32/32768/32768/manual_time       75.4 us          108 us         1857 bytes_per_second=623.26G/s 2D Schedule at 1/ Split block into y-dim/Vectorize, Factor: 2/Launch_Parameters[block(1/2/64)/grid(1/16384/1)/0]
Baseline_IndexSelect_Outer_fp32/128/16/manual_time                                                                         18.7 us         54.1 us         7477 bytes_per_second=1.22466G/s
Baseline_IndexSelect_Outer_fp32/512/16/manual_time                                                                         18.7 us         54.0 us         7475 bytes_per_second=1.22607G/s
Baseline_IndexSelect_Outer_fp32/4096/16/manual_time                                                                        18.6 us         53.9 us         7473 bytes_per_second=1.23084G/s
Baseline_IndexSelect_Outer_fp32/32768/16/manual_time                                                                       18.9 us         54.1 us         7418 bytes_per_second=1.21592G/s
...
Baseline_IndexSelect_Outer_fp32/128/32768/manual_time                                                                       186 us          222 us          753 bytes_per_second=252.216G/s
Baseline_IndexSelect_Outer_fp32/512/32768/manual_time                                                                       186 us          222 us          752 bytes_per_second=252.024G/s
Baseline_IndexSelect_Outer_fp32/4096/32768/manual_time                                                                      188 us          224 us          745 bytes_per_second=249.875G/s
Baseline_IndexSelect_Outer_fp32/32768/32768/manual_time                                                                     196 us          232 us          715 bytes_per_second=240.223G/s
```